WCM-43: Set up SQL isolation automatically via ZopeLayer
