+++ core/src/zeit/objectlog/testing.py
@@ -1,15 +1,13 @@
 import persistent
 
 import zeit.cms.testing
-import zeit.connector.testing
 
 
 ZCML_LAYER = zeit.cms.testing.ZCMLLayer(
     features=['zeit.connector.sql'],
-    bases=(zeit.cms.testing.CONFIG_LAYER, zeit.connector.testing.SQL_CONFIG_LAYER),
+    bases=(zeit.cms.testing.CONFIG_LAYER, zeit.cms.testing.SQL_LAYER),
 )
-SQL_LAYER = zeit.connector.testing.SQLIsolationLayer(ZCML_LAYER)
-ZOPE_LAYER = zeit.cms.testing.ZopeLayer(SQL_LAYER)
+ZOPE_LAYER = zeit.cms.testing.ZopeLayer(ZCML_LAYER)
 
 
 class FunctionalTestCase(zeit.cms.testing.FunctionalTestCase):
